26.1 Update Minecraft is now Unobfuscated and requires Java 25 Lots of changes/code cleanup. See https://github.com/MinecraftForge/MinecraftForge-Experimental/tree/26.1-dev for non-squashed commits.

+++ src/main/java/net/minecraftforge/event/entity/player/PlayerFlyableFallEvent.java
@@ -12,17 +12,16 @@ import org.jspecify.annotations.NullMarked;
 
 /**
  * Occurs when a player falls, but is able to fly.  Doesn't need to be cancelable, this is mainly for notification purposes.
- * @author Mithion
  */
 @NullMarked
 public final class PlayerFlyableFallEvent extends MutableEvent implements PlayerEvent {
     public static final EventBus<PlayerFlyableFallEvent> BUS = EventBus.create(PlayerFlyableFallEvent.class);
 
     private final Player player;
-    private float distance;
+    private double distance;
     private float multiplier;
 
-    public PlayerFlyableFallEvent(Player player, float distance, float multiplier) {
+    public PlayerFlyableFallEvent(Player player, double distance, float multiplier) {
         this.player = player;
         this.distance = distance;
         this.multiplier = multiplier;
@@ -33,7 +32,7 @@ public final class PlayerFlyableFallEvent extends MutableEvent implements Player
         return player;
     }
 
-    public float getDistance() { return distance;}
+    public double getDistance() { return distance;}
     public void setDistance(float distance) { this.distance = distance; }
     public float getMultiplier() { return multiplier; }
     public void setMultiplier(float multiplier) { this.multiplier = multiplier; }
